Add job and schedule policies. Add worker ID to job. Change permission assignment

+++ app/Http/Controllers/Api/JobScheduleController.php
@@ -6,6 +6,7 @@ use App\Http\Requests;
 use App\Http\Controllers\Controller;
 
 use App\Models\JobSchedule;
+use Illuminate\Auth\Access\AuthorizationException;
 use Illuminate\Http\Request;
 use Illuminate\Http\Response;
 
@@ -15,9 +16,11 @@ class JobScheduleController extends Controller
      * Display a listing of the resource.
      *
      * @return Response
+     * @throws AuthorizationException
      */
     public function index(Request $request)
     {
+        $this->authorize('viewAny');
         $jobschedule = JobSchedule::latest()->paginate(25);
 
         return $jobschedule;
@@ -29,10 +32,11 @@ class JobScheduleController extends Controller
      * @param Request $request
      *
      * @return Response
+     * @throws AuthorizationException
      */
     public function store(Request $request)
     {
-        
+        $this->authorize('create');
         $jobschedule = JobSchedule::create($request->all());
 
         return response()->json($jobschedule, 201);
@@ -41,13 +45,15 @@ class JobScheduleController extends Controller
     /**
      * Display the specified resource.
      *
-     * @param  int  $id
+     * @param int $id
      *
      * @return Response
+     * @throws AuthorizationException
      */
     public function show($id)
     {
         $jobschedule = JobSchedule::findOrFail($id);
+        $this->authorize('view', $jobschedule);
 
         return $jobschedule;
     }
@@ -56,14 +62,16 @@ class JobScheduleController extends Controller
      * Update the specified resource in storage.
      *
      * @param Request $request
-     * @param  int  $id
+     * @param int $id
      *
      * @return Response
+     * @throws AuthorizationException
      */
     public function update(Request $request, $id)
     {
-        
         $jobschedule = JobSchedule::findOrFail($id);
+        $this->authorize('edit', $jobschedule);
+
         $jobschedule->update($request->all());
 
         return response()->json($jobschedule, 200);
@@ -72,12 +80,16 @@ class JobScheduleController extends Controller
     /**
      * Remove the specified resource from storage.
      *
-     * @param  int  $id
+     * @param int $id
      *
      * @return Response
+     * @throws AuthorizationException
      */
     public function destroy($id)
     {
+        $jobschedule = JobSchedule::findOrFail($id);
+        $this->authorize('delete', $jobschedule);
+
         JobSchedule::destroy($id);
 
         return response()->json(null, 204);
